Fix vm_val_posr int8_t overflow when heap > 127 bytes
Heap byte offsets >= 128 were silently truncated to negative int8_t
values before being passed to vm_val_posr, causing the formula to
produce positive (invalid) slot numbers. This corrupted bytecode slot
assignments for any rule whose heap exceeded 127 bytes, leading to a
FATAL error at runtime.

+++ HeishaMon/src/rules/rules.cpp
@@ -1475,7 +1475,7 @@ static int16_t vm_val_pos(int8_t pos) {
   }
 }
 
-static int32_t vm_val_posr(int8_t pos) {
+static int32_t vm_val_posr(int16_t pos) {
   return ((pos-4)/rule_max_var_bytes()*-1)-1;
 }
